Add support for embedded documents
If an update or create contains an embedded document (eg
{name: 'John', address: {street: {...}}}, we need to include the embedded
address.street variable in the data payload. Prisma doesn't yet have a spec for this
(they're working on it at https://github.com/prisma/prisma/issues/2836), so in the
meantime, we just include it directly in the payload by processing the variables
recursively.