Allow React Admin to pass extra query variables which get forwarded to the server
With tasks, we want to allow a singleEvents flag to be passed along with the query. It's a top level query variable rather than a filter, so there's currently no way to pass it from the admin app into the data provider -- this provides a simple mechanism for doing so.